fix: address review comment 4888776879 — split attention events suite into parts
Move the 881-line monolith into test_workspace_attention_events_parts/
mirroring workspace_repo_attention coverage while preserving all 19 tests.
