+++ tests/unit/db/test_workspace_attention_events_parts/conftest.py
@@ -1,0 +1,16 @@
+"""Fixtures for workspace attention event repository tests."""
+
+from __future__ import annotations
+
+from collections.abc import AsyncIterator
+
+import pytest
+from sqlalchemy.ext.asyncio import AsyncSession
+
+from tests.postgres import postgres_test_session
+
+
+@pytest.fixture
+async def session() -> AsyncIterator[AsyncSession]:
+    async with postgres_test_session() as s:
+        yield s
